Puts reg. analysis methods in RegionStatus. Lists rooms on construction.
